+++ blog/src/app/login/page.tsx
@@ -1,97 +1,9 @@
-"use client";
+import { LoginCard } from "~/components/login/login-card";
 
-import { useState } from "react";
-import Link from "next/link";
-import { Button } from "~/components/ui/button";
-import { Input } from "~/components/ui/input";
-import { Label } from "~/components/ui/label";
-import {
-  Card,
-  CardContent,
-  CardDescription,
-  CardFooter,
-  CardHeader,
-  CardTitle,
-} from "~/components/ui/card";
-import type React from "react";
-import { useLogin } from "~/hooks/login-hook";
-
-interface LoginFormData {
-  Username: string;
-  Password: string;
-}
-
-export default function Login({
-  Username: defaultUsername = "",
-  Password: defaultPassword = "",
-}: Partial<LoginFormData>) {
-  const [Username, setUsername] = useState(defaultUsername);
-  const [Password, setPassword] = useState(defaultPassword);
-
-  const { login, isLoading } = useLogin();
-
-  const handleSubmit = async (e: React.FormEvent) => {
-    e.preventDefault();
-
-    login({
-      username: Username,
-      password: Password,
-    });
-  };
+export default function Login() {
   return (
     <div className="flex min-h-[calc(100vh-4rem)] items-center justify-center py-12">
-      <Card className="w-full max-w-md">
-        <CardHeader>
-          <CardTitle className="text-2xl">Log in</CardTitle>
-          <CardDescription>
-            Enter your username and password to access your account
-          </CardDescription>
-        </CardHeader>
-        <form onSubmit={handleSubmit}>
-          <CardContent className="space-y-4">
-            <div className="space-y-2">
-              <Label htmlFor="username">Username</Label>
-              <Input
-                id="username"
-                type="text"
-                placeholder="username"
-                value={Username}
-                onChange={(e) => setUsername(e.target.value)}
-                required
-              />
-            </div>
-            <div className="space-y-2">
-              <div className="flex items-center justify-between">
-                <Label htmlFor="password">Password</Label>
-                <Link
-                  href="/forgot-password"
-                  className="text-muted-foreground hover:text-primary text-xs"
-                >
-                  Forgot password?
-                </Link>
-              </div>
-              <Input
-                id="password"
-                type="password"
-                value={Password}
-                onChange={(e) => setPassword(e.target.value)}
-                required
-              />
-            </div>
-          </CardContent>
-          <CardFooter className="mt-4 flex flex-col space-y-4">
-            <Button type="submit" className="w-full" disabled={isLoading}>
-              {isLoading ? "Logging in..." : "Log in"}
-            </Button>
-            <p className="text-muted-foreground text-center text-sm">
-              Don&apos;t have an account?{" "}
-              <Link href="/register" className="text-primary hover:underline">
-                Sign up
-              </Link>
-            </p>
-          </CardFooter>
-        </form>
-      </Card>
+      <LoginCard />
     </div>
   );
 }
